Refactor into separate fs projects. (#67)
* Refactor into separate fs projects.

When device-scanner was first created, there was not great support for
multiple fsproj files. Namely, rollup could not build multiple input to
outputs, jest could not run multiple projects, and fable could not
update multipe fsproj over time easily.

Due to this, the existing version of device-scanner is a single fsproj
that contains all source and tests. This solved the previous issues, but
is not as nice from an organizational standpoint as different concerns
are merged together.

support for monorepo scenarios has improved recently; rollup can handle multiple input ->
outputs, jest can run multiple projects at once, and fable can update
multiple fsproj over time.

We are still missing the ability to run dotnet clitools at the solution
level, but that looks like it will included in an upcoming version of
paket.

With these changes, we can simplify building with rollup, encapsulate
testing concerns (integration tests can be totally separated from the
rest of the project, and be run as such), and allow for easier
publishing of projects that contain shared domain types.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* add more project to sln

Signed-off-by: Joe Grund <joe.grund@intel.com>